+++ src/axom/mint/examples/mint_unstructured_mixed_topology_mesh.cpp
@@ -19,7 +19,7 @@
 #include <random> /* for random number generator */
 
 using namespace axom;
-using axom::slic::UnitTestLogger;
+using axom::slic::SimpleLogger;
 
 inline bool appendQuad(axom::IndexType i, axom::IndexType j)
 {
@@ -29,7 +29,7 @@ inline bool appendQuad(axom::IndexType i, axom::IndexType j)
 //------------------------------------------------------------------------------
 int main(int AXOM_NOT_USED(argc), char** AXOM_NOT_USED(argv))
 {
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
 
   constexpr int DIMENSION = 2;
   constexpr axom::IndexType X_EXTENT = 11;